move clean clone into git pull as an option
