+++ lib/cmd/common/git.js
@@ -26,9 +26,12 @@ function git(argv, cb) {
   if (action === 'props') {
     appId = fhc.appId(args[1]);
     return props(appId, cb);
-  } else if (action === 'pull') {
+  } else if (action === 'pull' && !argv.clean) {
     appId = fhc.appId(args[1]);
-    return pull(appId, cb);
+    return pull(appId, argv.clean, cb);
+  } else if (action === 'pull' && argv.clean) {
+    appId = fhc.appId(args[1]);
+    return cleanClone(appId, cb);
   } else if (action === 'set') {
     args.shift();
     if (args.length === 0 || args.length > 3) return cb(git.usage);
@@ -92,6 +95,34 @@ function set(appId, k, v, cb) {
           return cb(remoteData.messsage, remoteData);
         }
         return cb(undefined, remoteData);
+      }
+    });
+  });
+}
+
+function cleanClone (appId, cb) {
+  var payload = {
+    clean: true
+  };
+
+  if (!appId) {
+    return cb(
+      new Error(i18n._('appId is required to perform a clean clone'))
+    );
+  }
+
+  millicore.widgForAppId(appId, function (err, widgId) {
+    if (err) {
+      return cb(err);
+    }
+    //Sample URL: https://domain.feedhenry.com/box/api/connectors/{widget}/apps/{appId}/pull
+    fhreq.POST(fhreq.getFeedHenryUrl(), '/box/api/projects/' + widgId + '/apps/' + appId + '/pull', payload, function (err, data) {
+      if (err) {
+        return cb(err);
+      } else if (!data.cacheKeys) {
+        return cb(i18n._('Error in clean clone, unexpected response format'));
+      } else {
+        return cb(null, i18n._('clean clone successful'));
       }
     });
   });
